+++ gsi-server/src/main/java/de/gematik/idp/gsi/server/data/GsiConstants.java
@@ -16,6 +16,10 @@
 
 package de.gematik.idp.gsi.server.data;
 
+import de.gematik.idp.field.ClaimName;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
 import java.util.Set;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
@@ -38,4 +42,22 @@ public final class GsiConstants {
 
   public static final String FEDIDP_PAR_AUTH_ENDPOINT = "/PAR_Auth";
   public static final String FED_SIGNED_JWKS_ENDPOINT = "/jws.json";
+  public static final Map<String, List<ClaimName>> SCOPES_TO_CLAIM_MAP = new HashMap<>();
+
+  static {
+    SCOPES_TO_CLAIM_MAP.put("openid", List.of());
+    SCOPES_TO_CLAIM_MAP.put("urn:telematik:geburtsdatum", List.of(ClaimName.BIRTHDATE));
+    SCOPES_TO_CLAIM_MAP.put("urn:telematik:alter", List.of(ClaimName.TELEMATIK_ALTER));
+    SCOPES_TO_CLAIM_MAP.put(
+        "urn:telematik:display_name", List.of(ClaimName.TELEMATIK_DISPLAY_NAME));
+    SCOPES_TO_CLAIM_MAP.put("urn:telematik:given_name", List.of(ClaimName.TELEMATIK_GIVEN_NAME));
+    SCOPES_TO_CLAIM_MAP.put("urn:telematik:geschlecht", List.of(ClaimName.TELEMATIK_GESCHLECHT));
+    SCOPES_TO_CLAIM_MAP.put("urn:telematik:email", List.of(ClaimName.TELEMATIK_EMAIL));
+    SCOPES_TO_CLAIM_MAP.put(
+        "urn:telematik:versicherter",
+        List.of(
+            ClaimName.TELEMATIK_PROFESSION,
+            ClaimName.TELEMATIK_ID,
+            ClaimName.TELEMATIK_ORGANIZATION));
+  }
 }