+++ gsi-server/src/main/java/de/gematik/idp/gsi/server/data/Metadata.java
@@ -18,16 +18,8 @@ package de.gematik.idp.gsi.server.data;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.AllArgsConstructor;
 import lombok.Builder;
-import lombok.Getter;
 
-@Getter
 @Builder
-@AllArgsConstructor
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class Metadata {
-
-  private OpenidProvider openidProvider;
-  private FederationEntity federationEntity;
-}
+public record Metadata(OpenidProvider openidProvider, FederationEntity federationEntity) {}