+++ gsi-server/src/main/java/de/gematik/idp/gsi/server/data/QRCodeGenerator.java
@@ -1,0 +1,47 @@
+/*
+ *  Copyright 2023 gematik GmbH
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package de.gematik.idp.gsi.server.data;
+
+import com.google.zxing.BarcodeFormat;
+import com.google.zxing.EncodeHintType;
+import com.google.zxing.MultiFormatWriter;
+import com.google.zxing.client.j2se.MatrixToImageWriter;
+import com.google.zxing.common.BitMatrix;
+import com.google.zxing.qrcode.decoder.ErrorCorrectionLevel;
+import java.io.ByteArrayOutputStream;
+import java.util.Base64;
+import java.util.HashMap;
+import java.util.Map;
+import lombok.SneakyThrows;
+
+public abstract class QRCodeGenerator {
+
+  @SneakyThrows
+  public static String generate(final String content) {
+    final Map<EncodeHintType, Object> hints = new HashMap<>();
+    hints.put(EncodeHintType.ERROR_CORRECTION, ErrorCorrectionLevel.H);
+    hints.put(EncodeHintType.CHARACTER_SET, "UTF-8");
+
+    final BitMatrix matrix =
+        new MultiFormatWriter().encode(content, BarcodeFormat.QR_CODE, 200, 200, hints);
+    final ByteArrayOutputStream baos = new ByteArrayOutputStream();
+    MatrixToImageWriter.writeToStream(matrix, "PNG", baos);
+    final byte[] imageBytes = baos.toByteArray();
+    final String base64Image = Base64.getEncoder().encodeToString(imageBytes);
+    return "data:image/jpeg;base64," + base64Image;
+  }
+}